backend/requirements.txt
django>=5.2.1
djangorestframework>=3.16.0
channels>=4.2.2
channels-redis>=4.2.1
ffmpeg-python>=0.2.0
python-dotenv>=1.1.0
redis>=6.1.0 